==> nbs/api/77_plot.ipynb <==
# plot

> Conc plot result class.
- toc: false
- page-layout: full

#| default_exp plot

#| hide
from nbdev.showdoc import *

#| export
from __future__ import annotations
from fastcore.basics import patch
from IPython.display import display, HTML

## Using the Plot class

Conc plotting functionality outputs a `Plot` object. This gives flexibility to display the plot in a notebook, write it to file, or return the HTML for the result.

## Plot class API reference

#| export
class Plot:
	def __init__(self,
			  type: str, # report type
			  html: str, # HTML representation of the plot, may include js/css
			  ): 

		""" Class for handling delivery of Conc plot results """
		self.type = type
		self.html = html
		


#| export
@patch
def to_html(self: Plot
			 ):
	""" Return HTML string for plot """
	
	return self.html

#| export
@patch
def to_file(self: Plot,
			filename:str, # filename to write HTML to
			encoding = 'utf-8', # encoding string for writing file 
			):
	""" Return HTML string for plot """

	with open(filename, 'w', encoding=encoding) as f:
		f.write(self.html)

#| export
@patch
def display(self: Plot
			   ):
	""" Show the plot in a Jupyter notebook """
	
	display(HTML(self.html))


#| hide
import nbdev; nbdev.nbdev_export()
